+++ examples/demo/.umirc.ts
@@ -2,6 +2,7 @@ import { defineConfig } from 'umi';
 import path from 'path';
 
 export default defineConfig({
+  fastRefresh: {},
   plugins: [path.join(__dirname, '../../lib')],
   nodeModulesTransform: {
     type: 'none',