+++ examples/main+renderer/src/renderer/.umirc.ts
@@ -3,6 +3,7 @@ import { Configuration } from 'webpack';
 import { resolve } from 'path';
 
 export default defineConfig({
+  fastRefresh: {},
   routes: [{ path: '/', component: '@/pages/index' }],
   plugins: [resolve(__dirname, '../../../../lib')],
   alias: {